Reorganize project structure and take into account savings account

+++ src/Model/BankAccount.php
@@ -22,7 +22,7 @@ class BankAccount
 
     public float $balance;
 
-    public ?int $coming;
+    public ?float $coming;
 
     public bool $display;
